+++ CHANGELOG.md
@@ -9,3 +9,8 @@
 ## 1.0.2
 
 - Hotfix command line changes from flutter to dart
+
+## 2.0.0
+
+- Re-designed the package to be a global executable instead
+- Implement better parsing and support case for overridden dependencies